[titusan] creating a way for users to edit posts
